requirements.txt
streamlit
sentence-transformers
pymupdf
python-docx
numpy
scikit-learn
